feat 회원정보 수정 및 탈퇴 추가

+++ src/main/java/com/petmily/petmily/User/User.java
@@ -16,7 +16,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "userkey")
-    private int userkey;
+    private Long userkey;
 
     @Column(name = "userid")
     private String userid;
